requirements.txt
google-adk>=1.0.0
mcp
arxiv
python-dotenv
